% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edit_Figures.R
\name{Edit_Figures}
\alias{Edit_Figures}
\title{Edit Color of Cartoon Figures}
\usage{
Edit_Figures()
}
\value{
A Shiny app allowing for character editing and downloading
}
\description{
Edit Color of Cartoon Figures
}
\examples{
Edit_Figures()
}
