% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{figure_information}
\alias{figure_information}
\title{Information on Figure Characteristics and Default Colors}
\format{
\subsection{\code{figure_information}}{

A data frame with 92 rows and 4 columns:
\describe{
\item{Part}{part shown in svg - head or clothes}
\item{Label}{label shown in shiny app dropdown menu}
\item{Item}{object whose color is specified}
\item{Color}{color of Item specified}
...
}
}
}
\source{
created from the default cartoons
}
\usage{
figure_information
}
\description{
Information on Figure Characteristics and Default Colors
}
\keyword{datasets}
