% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_Testimony.R
\name{Format_Testimony}
\alias{Format_Testimony}
\title{Format Testimony}
\usage{
Format_Testimony(testimony_file_path)
}
\arguments{
\item{testimony_file_path}{path to testimony csv file}
}
\value{
a formatted csv file with appropriate div classes for speech bubbles
}
\description{
Format Testimony
}
