% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_Template.R
\name{Save_Template}
\alias{Save_Template}
\title{Save RShiny Courtroom Study Template}
\usage{
Save_Template(file_path, overwrite = FALSE)
}
\arguments{
\item{file_path}{path where template should be saved}

\item{overwrite}{TRUE to overwrite existing files}
}
\value{
RShiny app template will be saved in designated location
}
\description{
Save RShiny Courtroom Study Template
}
